added support of filtering instances by tags

+++ ec2_instance_clean.py
@@ -10,7 +10,8 @@ import logging
 import argparse
 
 parser = argparse.ArgumentParser('To list/clean up instances cross regions')
-parser.add_argument('--key_name', dest='key_name',action='store',help='specify for owner, seperated by ","',required=True)
+parser.add_argument('--key_name', dest='key_name',action='store',help='specify for owner, seperated by ","',required=False)
+parser.add_argument('--tags', dest='tags',action='store',help='specify for tags, seperated by ","',required=False)
 parser.add_argument('--delete', dest='delete',action='store_true',help='optional, specify for delete instances, otherwise list only',required=False)
 parser.add_argument('-d', dest='is_debug',action='store_true',help='optional, run in debug mode', required=False,default=False)
 parser.add_argument('--skip_region', dest='skip_region',action='store',help='optional skip regions, seperated by ","',required=False,default=None)
@@ -45,15 +46,29 @@ for region in region_list['Regions']:
     for instance in s['Reservations']:
         #log.info(instance)
         try:
-            for key_name in args.key_name.split(','):
-                if key_name in instance['Instances'][0]['KeyName']:
-                    instance_id = instance['Instances'][0]['InstanceId']
-                    log.info('%s %s' % (instance['Instances'][0]['KeyName'],instance_id ))
-                    if args.delete:
-                        ec2 = boto3.resource('ec2',region_name=region_name)
-                        vm = ec2.Instance(instance_id)
-                        vm.terminate()
-                        log.info('%s terminated' % instance_id)
+            if args.key_name is not None:
+                for key_name in args.key_name.split(','):
+                    if instance['Instances'][0].has_key('KeyName'):
+                        if key_name in instance['Instances'][0]['KeyName']:
+                            instance_id = instance['Instances'][0]['InstanceId']
+                            log.info('Key:%s id:%s' % (instance['Instances'][0]['KeyName'],instance_id ))
+                            if args.delete:
+                                ec2 = boto3.resource('ec2',region_name=region_name)
+                                vm = ec2.Instance(instance_id)
+                                vm.terminate()
+                                log.info('%s terminated' % instance_id)
+            if args.tags is not None:
+                for tag in args.tags.split(','):
+                    if instance['Instances'][0].has_key('Tags'):
+                        for val  in instance['Instances'][0]['Tags'][0].values():
+                            if tag in val:
+                                instance_id = instance['Instances'][0]['InstanceId']
+                                log.info('tag:%s id:%s' % (instance['Instances'][0]['Tags'][0]['Value'],instance_id ))
+                                if args.delete:
+                                    ec2 = boto3.resource('ec2',region_name=region_name)
+                                    vm = ec2.Instance(instance_id)
+                                    vm.terminate()
+                                    log.info('%s terminated' % instance_id)
         except KeyError as err:
-            log.info("No key found")
+            log.info("No such words found %s" % err)
